Golang Build: add the Run variant to the build system
This change alters how the go subprocess is killed to ensure that
the temporary binary executed during "go run" is also killed so
that users can cancel the Run variant like the other variants.

Additionally, there is support for a file path to the run variant
so that users can create a project setting that specifies the file
to always pass to "go run". The file path may be relative to
$GOPATH/src/, for multi-user or cross-platform development.
The relative file path functionality supports multiple entries in
$GOPATH.

Change-Id: I7cf3b3e9ff89665ea0cd776fe6facb164575fecb
Reviewed-on: https://go-review.googlesource.com/16441
Reviewed-by: Glenn Lewis <gmlewis@google.com>
Reviewed-by: Jason Buberel <jbuberel@google.com>